[FIX] stock: move without initial demand/quantity
When processing picking containing moves without
quantity done and without initial demand. These moves
were moving automatically to a new backorder.

It was also possible to validate picking without
quantity done for all moves.

This commit raise an error if the user try to valiate
a picking with all moves without quantity done and without
initial demand. Or cancel all these moves if the picking
is validate with correct moves.