[FIX] web: restore hovering effects on folded kanban columns
There are two kind of hovering effects:

(1) When the user hovers the column with the mouse

It was broken with https://github.com/odoo/odoo/commit/cb5fdbfa60677296b27afef5fb1db031708ea705
because the "o_kanban_header" class was renamed to
"o_kanban_header_title" and the effect still applied on the
"o_kanban_header" element.

(2) When the user is about to drop a kanban record

It was broken with https://github.com/odoo/odoo/commit/acabe00081e021935fcb692049b83b0cf745a938
because styling the "o_kanban_hover" class was removed by mistake.