[FIX] mrp: check that a SN as not already been used
When you are processing a MO or WO, you can set two times the same serial
number for the same product.

To avoid this behavior, we created a onchange on the stock move line to
warn the user when he enter two times the same SN. We also create a
constraint on stock move line to avoid having two times same SN in a
picking.