[FIX] base: tests, reflect unaccent behavior of d03f39ffb8d16a5e7bbf2b8bf22bdb8781972fb3
Since the revision
d03f39ffb8d16a5e7bbf2b8bf22bdb8781972fb3

The `::text` is included within the `unaccent` call
The expession tests have never been changed according
to this change of behavior in the building of the unaccented
queries. This has never been detected by runbot because
you need to start the Odoo server with the option
`--unaccent`
and to create the database with the unaccent extension
`CREATE EXTENSION unaccent`
to be able to fail the former tests,
and this is not the case on runbot.