[FIX] web: allow creating new kanban columns with progressbar
Before this fix, trying to create a column in a kanban view where there
were progressbars was opening a traceback. This was because the newly
create column is made from data which are guessed in the client (no RPC
is done), so this guess had to be adapted to progressbars.