[FIX] web: focus issues after onchange in x2many with buttons
https://github.com/odoo/odoo/commit/7bf2aa3299db5cb1188a4c46ddef6d1c81936c6f
fixes most post-onchange focus issues for x2many fields, but does not
account for "button columns" properly. More precisely, the current
widget is computed via a position index taking buttons into account in
an array that ignores them. This caused the index to be out of bounds,
resulting in a traceback. This commit fixes this issue.